# elpe/ui/sampleResponseAndSwagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Risk analysis REST API",
        "version": "1.0.0",
        "title": "We metrix Risk analysis API",
        "contact": {
            "name": "John Garos, head of big data",
            "url": "https://wemetrix.com/",
            "email": "info@wemetrix.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        }
    },
    "host": "localhost:8090",
    "basePath": "/",
...
    "paths": {
        "/repository/clienteledistribution": {
            "get": {
                "tags": [
                    "business-controller"
                ],
                "summary": "clienteleDistribution",
                "operationId": "clienteleDistributionUsingGET",
                "produces": [
                    "*/*"
                ],
                "parameters": [
                    {
                        "name": "authenticated",
                        "in": "query",
                        "required": false,
                        "type": "boolean"
                    },
                    {
                        "name": "authorities[0].authority",
                        "in": "query",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "comparisonPeriod",
                        "in": "query",
                        "description": "comparisonPeriod",
                        "required": true,
                        "type": "integer",
                        "format": "int32"
                    },
                    {
                        "name": "contextId",
                        "in": "query",
                        "description": "contextId",
                        "required": true,
                        "type": "integer",
                        "format": "int32"
                    },
                    {
                        "name": "credentials",
                        "in": "query",
                        "required": false,
                        "type": "object"
                    },
                    {
                        "name": "details",
                        "in": "query",
                        "required": false,
                        "type": "object"
                    },
                    {
                        "name": "portfolioId",
                        "in": "query",
                        "description": "portfolioId",
                        "required": true,
                        "type": "integer",
                        "format": "int32"
                    },
                    {
                        "name": "principal",
                        "in": "query",
                        "required": false,
                        "type": "object"
                    },
                    {
                        "name": "snapshotDate",
                        "in": "query",
                        "description": "snapshotDate",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ClienteleDistributionDto"
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                },
                "deprecated": false
            }
        },
        "/repository/clientelestatistics": {
            "get": {
                "tags": [
                    "business-controller"
                ],
                "summary": "clienteleStatistics",
                "operationId": "clienteleStatisticsUsingGET",
                "produces": [
                    "*/*"
                ],
                "parameters": [
                    {
                        "name": "authenticated",
                        "in": "query",
                        "required": false,
                        "type": "boolean"
                    },
                    {
                        "name": "authorities[0].authority",
                        "in": "query",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "comparisonPeriod",
                        "in": "query",
                        "description": "comparisonPeriod",
                        "required": true,
                        "type": "integer",
                        "format": "int32"
                    },
                    {
                        "name": "contextId",
                        "in": "query",
                        "description": "contextId",
                        "required": true,
                        "type": "integer",
                        "format": "int32"
                    },
                    {
                        "name": "credentials",
                        "in": "query",
                        "required": false,
                        "type": "object"
                    },
                    {
                        "name": "details",
                        "in": "query",
                        "required": false,
                        "type": "object"
                    },
                    {
                        "name": "portfolioId",
                        "in": "query",
                        "description": "portfolioId",
                        "required": true,
                        "type": "integer",
                        "format": "int32"
                    },
                    {
                        "name": "principal",
                        "in": "query",
                        "required": false,
                        "type": "object"
                    },
                    {
                        "name": "snapshotDate",
                        "in": "query",
                        "description": "snapshotDate",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ClienteleStatisticsDto"
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                },
                "deprecated": false
            }
        },
        "/repository/clientelestatisticsgraph": {
            "get": {
                "tags": [
                    "business-controller"
                ],
                "summary": "clienteleStatisticsGraph",
                "operationId": "clienteleStatisticsGraphUsingGET",
                "produces": [
                    "*/*"
                ],
                "parameters": [
                    {
                        "name": "authenticated",
                        "in": "query",
                        "required": false,
                        "type": "boolean"
                    },
                    {
                        "name": "authorities[0].authority",
                        "in": "query",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "comparisonPeriod",
                        "in": "query",
                        "description": "comparisonPeriod",
                        "required": true,
                        "type": "integer",
                        "format": "int32"
                    },
                    {
                        "name": "contextId",
                        "in": "query",
                        "description": "contextId",
                        "required": true,
                        "type": "integer",
                        "format": "int32"
                    },
                    {
                        "name": "credentials",
                        "in": "query",
                        "required": false,
                        "type": "object"
                    },
                    {
                        "name": "details",
                        "in": "query",
                        "required": false,
                        "type": "object"
                    },
                    {
                        "name": "portfolioId",
                        "in": "query",
                        "description": "portfolioId",
                        "required": true,
                        "type": "integer",
                        "format": "int32"
                    },
                    {
                        "name": "principal",
                        "in": "query",
                        "required": false,
                        "type": "object"
                    },
                    {
                        "name": "snapshotDate",
                        "in": "query",
                        "description": "snapshotDate",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ClienteleStatisticsGraphDto"
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                },
                "deprecated": false
            }
        },
        "/repository/configinfo": {
            "get": {
                "tags": [
                    "business-controller"
                ],
                "summary": "Retrieves the configuration info",
                "operationId": "getConfigInfoUsingGET",
                "produces": [
                    "*/*"
                ],
                "parameters": [
                    {
                        "name": "authenticated",
                        "in": "query",
                        "required": false,
                        "type": "boolean"
                    },
                    {
                        "name": "authorities[0].authority",
                        "in": "query",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "credentials",
                        "in": "query",
                        "required": false,
                        "type": "object"
                    },
                    {
                        "name": "details",
                        "in": "query",
                        "required": false,
                        "type": "object"
                    },
                    {
                        "name": "principal",
                        "in": "query",
                        "required": false,
                        "type": "object"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ConfigInfo"
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                },
                "deprecated": false
            }
        },
        "/repository/contextinfo": {
            "get": {
                "tags": [
                    "business-controller"
                ],
                "summary": "Retrieves the context info",
                "operationId": "getContextInfoUsingGET",
                "produces": [
                    "*/*"
                ],
                "parameters": [
                    {
                        "name": "authenticated",
                        "in": "query",
                        "required": false,
                        "type": "boolean"
                    },
                    {
                        "name": "authorities[0].authority",
                        "in": "query",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "contextId",
                        "in": "query",
                        "description": "contextId",
                        "required": true,
                        "type": "integer",
                        "format": "int32"
                    },
                    {
                        "name": "credentials",
                        "in": "query",
                        "required": false,
                        "type": "object"
                    },
                    {
                        "name": "details",
                        "in": "query",
                        "required": false,
                        "type": "object"
                    },
                    {
                        "name": "principal",
                        "in": "query",
                        "required": false,
                        "type": "object"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ContextInfo"
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                },
                "deprecated": false
            }
        },
        "/repository/customers": {
            "get": {
                "tags": [
                    "business-controller"
                ],
                "summary": "customers",
                "operationId": "customersUsingGET",
                "produces": [
                    "*/*"
                ],
                "parameters": [
                    {
                        "name": "authenticated",
                        "in": "query",
                        "required": false,
                        "type": "boolean"
                    },
                    {
                        "name": "authorities[0].authority",
                        "in": "query",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "contextId",
                        "in": "query",
                        "description": "contextId",
                        "required": true,
                        "type": "integer",
                        "format": "int32"
                    },
                    {
                        "name": "credentials",
                        "in": "query",
                        "required": false,
                        "type": "object"
                    },
                    {
                        "name": "details",
                        "in": "query",
                        "required": false,
                        "type": "object"
                    },
                    {
                        "name": "length",
                        "in": "query",
                        "description": "length",
                        "required": true,
                        "type": "integer",
                        "format": "int32"
                    },
                    {
                        "name": "offset",
                        "in": "query",
                        "description": "offset",
                        "required": true,
                        "type": "integer",
                        "format": "int32"
                    },
                    {
                        "name": "portfolioId",
                        "in": "query",
                        "description": "portfolioId",
                        "required": true,
                        "type": "integer",
                        "format": "int32"
                    },
                    {
                        "name": "principal",
                        "in": "query",
                        "required": false,
                        "type": "object"
                    },
                    {
                        "name": "scenarioId",
                        "in": "query",
                        "description": "scenarioId",
                        "required": true,
                        "type": "integer",
                        "format": "int32"
                    },
                    {
                        "name": "snapshotDate",
                        "in": "query",
                        "description": "snapshotDate",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "sortColumn",
                        "in": "query",
                        "description": "sortColumn",
                        "required": true,
                        "type": "integer",
                        "format": "int32"
                    },
                    {
                        "name": "sortDescending",
                        "in": "query",
                        "description": "sortDescending",
                        "required": true,
                        "type": "boolean"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/CustomerListDto"
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                },
                "deprecated": false
            }
        },
...
    },
    "definitions": {
...
        "ClienteleDistributionDto": {
            "type": "object",
            "properties": {
                "clienteleDistributionList": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ClienteleDistributionItemCollectionDto"
                    }
                }
            },
            "title": "ClienteleDistributionDto"
        },
...
        "ClienteleStatisticsDto": {
            "type": "object",
            "properties": {
                "clientStatisticItemDtos": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ClientStatisticItemDto"
                    }
                },
                "lastPeriod": {
                    "type": "string"
                },
                "previousPeriod": {
                    "type": "string"
                }
            },
            "title": "ClienteleStatisticsDto"
        },
        "ClienteleStatisticsGraphDto": {
            "type": "object",
            "properties": {
                "lineDataPointCollection": {
                    "$ref": "#/definitions/LineGraphValuesCollection"
                }
            },
            "title": "ClienteleStatisticsGraphDto"
        },
...
        "ConfigInfo": {
            "type": "object",
            "properties": {
                "changePwRequired": {
                    "type": "boolean"
                },
                "contexts": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "copyright": {
                    "type": "string"
                },
                "fullName": {
                    "type": "string"
                },
                "helpFiles": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "loginName": {
                    "type": "string"
                },
                "organization": {
                    "type": "string"
                },
                "publicIp": {
                    "type": "string"
                },
                "pwRegEx": {
                    "type": "string"
                },
                "roles": {
                    "type": "array",
                    "items": {
                        "type": "integer",
                        "format": "int32"
                    }
                },
                "version": {
                    "type": "string"
                }
            },
            "title": "ConfigInfo"
        },
...
        "CustomerListDto": {
            "type": "object",
            "properties": {
                "customers": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/Customer"
                    }
                },
                "customersCount": {
                    "type": "integer",
                    "format": "int32"
                }
            },
            "title": "CustomerListDto"
        },
...
    }
}